+++ src/main/resources/client-api-config.yml
@@ -257,11 +257,11 @@ api:
               description: 결제 상태 (PENDING)
 
     confirm-payment:
-      url:
-      method:
+      url: /api/payments/{paymentId}
+      method: POST
       description: 결제 확정 (PortOne 결제 검증 및 주문 상태 업데이트)
       pathParams:
-        - name:
+        - name: paymentId
           description: 결제 ID
       response:
         body:
